requirements.txt
# Versão final do projeto CuidarBot para Imersão IA Alura+Google
Flask
opencv-python
pytesseract
Pillow
google-generativeai
python-dotenv
requests # Pywhatkit pode precisar, ou para outras chamadas HTTP
# pywhatkit # Se for tentar usar, mas lembre-se dos desafios
qrcode[pil]
